requirements.txt
streamlit==1.50.0
boto3==1.40.44
langchain==0.3.27
langchain-community==0.3.30
langchain-huggingface==0.3.1
pypdf==6.1.1
qdrant-client==1.15.1
sentence-transformers==5.1.1
python-dotenv==1.0.0
